pyproject.toml
[project]
name = "trading23-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "mplfinance>=0.12.10b0",
    "pandas-stubs>=2.2.3.241009",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "polars>=1.12.0",
    "psutil>=6.1.0",
    "pyarrow>=18.0.0",
    "pydantic>=2.9.2",
    "requests>=2.32.3",
    "scikit-learn>=1.5.2",
    "timm>=1.0.11",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.0",
    "typer>=0.13.0",
    "types-psutil>=6.1.0.20241102",
    "types-requests>=2.32.0.20241016",
    "types-tqdm>=4.67.0.20241119",
]
